fix: resolve remaining H-IGH/MEDIUM audit findings
LEAVE RACE CONDITION (HIGH):
- Pessimistic write lock on leave approval to prevent over-approval
- Concurrent approvals now serialize on DB row lock

PAYROLL DUPLICATES (HIGH):
- Skip employees who already have payroll for the period in batchCalculate

INFRASTRUCTURE (HIGH):
- Structured logging with logback-spring.xml (JSON for production)
- Prometheus metrics endpoint enabled
- Graceful shutdown configured
- H2 scoped to test only (security)
- Java version downgraded to 21 LTS (production stability)

CSV INJECTION (MEDIUM):
- ReportService sanitizes cells starting with =, +, -, @

MONEY TYPE FIX (MEDIUM):
- All Doubles for amount/hours fields converted to BigDecimal
- CashCollectionRequest, AdminCashCollectionRequest, SalaryAdvanceRequest,
  ManualAdvanceRequest, DailyLogRequest, CashCollectionReviewRequest

SILENT SWALLOWING FIX (MEDIUM):
- All 5 batch delete methods now log failures instead of empty catch blocks

SERVICE LAYER (MEDIUM):
- DashboardController no longer injects repositories directly
- New DashboardService extracted

PWA/TOKEN SECURITY:
- vite-plugin-pwa removed (was misconfigured)
- OAuth2 URL token params removed (token now in HttpOnly cookie only)

+++ allinone-backend/src/main/java/com/workflow/dto/CashCollectionReviewRequest.java
@@ -1,11 +1,12 @@
 package com.workflow.dto;
 
 import jakarta.validation.constraints.NotBlank;
+import java.math.BigDecimal;
 
 public class CashCollectionReviewRequest {
     @NotBlank private String status;
     private String reviewComment;
-    private Double amount;
+    private BigDecimal amount;
     private String customerName;
     private String customerPhone;
     private String customerAddress;
@@ -17,7 +18,7 @@ public class CashCollectionReviewRequest {
 
     public String getStatus() { return status; } public void setStatus(String status) { this.status = status; }
     public String getReviewComment() { return reviewComment; } public void setReviewComment(String reviewComment) { this.reviewComment = reviewComment; }
-    public Double getAmount() { return amount; } public void setAmount(Double amount) { this.amount = amount; }
+    public BigDecimal getAmount() { return amount; } public void setAmount(BigDecimal amount) { this.amount = amount; }
     public String getCustomerName() { return customerName; } public void setCustomerName(String customerName) { this.customerName = customerName; }
     public String getCustomerPhone() { return customerPhone; } public void setCustomerPhone(String customerPhone) { this.customerPhone = customerPhone; }
     public String getCustomerAddress() { return customerAddress; } public void setCustomerAddress(String customerAddress) { this.customerAddress = customerAddress; }
