+++ allinone-backend/src/main/java/com/workflow/dto/DailyLogRequest.java
@@ -1,11 +1,12 @@
 package com.workflow.dto;
 
 import jakarta.validation.constraints.NotBlank;
+import java.math.BigDecimal;
 import java.util.List;
 
 public class DailyLogRequest {
     private Long employeeId; private Long branchId; private String logDate;
-    private String startTime; private String endTime; private Double hoursWorked;
+    private String startTime; private String endTime; private BigDecimal hoursWorked;
     private String category; private String locationDescription;
     private Double locationLat; private Double locationLng;
     @NotBlank private String workDescription; private List<String> photoUrls; private Long assignedTaskId;
@@ -17,7 +18,7 @@ public class DailyLogRequest {
     public String getLogDate() { return logDate; } public void setLogDate(String logDate) { this.logDate = logDate; }
     public String getStartTime() { return startTime; } public void setStartTime(String startTime) { this.startTime = startTime; }
     public String getEndTime() { return endTime; } public void setEndTime(String endTime) { this.endTime = endTime; }
-    public Double getHoursWorked() { return hoursWorked; } public void setHoursWorked(Double hoursWorked) { this.hoursWorked = hoursWorked; }
+    public BigDecimal getHoursWorked() { return hoursWorked; } public void setHoursWorked(BigDecimal hoursWorked) { this.hoursWorked = hoursWorked; }
     public String getCategory() { return category; } public void setCategory(String category) { this.category = category; }
     public String getLocationDescription() { return locationDescription; }
     public void setLocationDescription(String locationDescription) { this.locationDescription = locationDescription; }
