+++ allinone-backend/src/main/java/com/workflow/dto/ManualAdvanceRequest.java
@@ -2,18 +2,19 @@ package com.workflow.dto;
 
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
+import java.math.BigDecimal;
 
 public class ManualAdvanceRequest {
     @NotNull private Long employeeId;
-    @NotNull @Positive private Double amount;
+    @NotNull @Positive private BigDecimal amount;
     private String reason;
 
     public ManualAdvanceRequest() {}
 
     public Long getEmployeeId() { return employeeId; }
     public void setEmployeeId(Long employeeId) { this.employeeId = employeeId; }
-    public Double getAmount() { return amount; }
-    public void setAmount(Double amount) { this.amount = amount; }
+    public BigDecimal getAmount() { return amount; }
+    public void setAmount(BigDecimal amount) { this.amount = amount; }
     public String getReason() { return reason; }
     public void setReason(String reason) { this.reason = reason; }
 }
