+++ allinone-backend/src/main/java/com/workflow/dto/SalaryAdvanceRequest.java
@@ -2,13 +2,14 @@ package com.workflow.dto;
 
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
+import java.math.BigDecimal;
 
 public class SalaryAdvanceRequest {
-    @NotNull @Positive private Double amount;
+    @NotNull @Positive private BigDecimal amount;
     private String reason;
 
     public SalaryAdvanceRequest() {}
 
-    public Double getAmount() { return amount; } public void setAmount(Double amount) { this.amount = amount; }
+    public BigDecimal getAmount() { return amount; } public void setAmount(BigDecimal amount) { this.amount = amount; }
     public String getReason() { return reason; } public void setReason(String reason) { this.reason = reason; }
 }
